split legend component into multiple component files
