+++ frontend/src/lib/components/InsightLegend/InsightLegend.tsx
@@ -1,19 +1,11 @@
 import './InsightLegend.scss'
-import { Button } from 'antd'
 import { useActions, useValues } from 'kea'
-import { IconLegend } from 'lib/lemon-ui/icons'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { trendsLogic } from 'scenes/trends/trendsLogic'
-import { InsightLabel } from 'lib/components/InsightLabel'
-import { getSeriesColor } from 'lib/colors'
-import { LemonCheckbox } from 'lib/lemon-ui/LemonCheckbox'
-import { formatCompareLabel } from 'scenes/insights/views/InsightsTable/columns/SeriesColumn'
 import { ChartDisplayType, FilterType, InsightType } from '~/types'
 import clsx from 'clsx'
-import { formatAggregationAxisValue } from 'scenes/insights/aggregationAxisFormat'
-import { IndexedTrendResult } from 'scenes/trends/types'
-import { useEffect, useRef } from 'react'
-import { isFilterWithDisplay, isTrendsFilter } from 'scenes/insights/sharedUtils'
+import { isFilterWithDisplay } from 'scenes/insights/sharedUtils'
+import { InsightLegendRow } from './InsightLegendRow'
 
 export interface InsightLegendProps {
     readOnly?: boolean
@@ -30,23 +22,11 @@ const trendTypeCanShowLegendDenyList = [
 
 const insightViewCanShowLegendAllowList = [InsightType.TRENDS, InsightType.STICKINESS]
 
-const shouldShowLegend = (filters: Partial<FilterType>): boolean =>
+export const shouldShowLegend = (filters: Partial<FilterType>): boolean =>
     insightViewCanShowLegendAllowList.includes(filters.insight || InsightType.TRENDS) &&
     isFilterWithDisplay(filters) &&
     !!filters.display &&
     !trendTypeCanShowLegendDenyList.includes(filters.display)
-
-export function InsightLegendButton(): JSX.Element | null {
-    const { filters } = useValues(insightLogic)
-    const { toggleInsightLegend } = useActions(insightLogic)
-
-    return shouldShowLegend(filters) && isFilterWithDisplay(filters) ? (
-        <Button className="InsightLegendButton" onClick={toggleInsightLegend}>
-            <IconLegend />
-            <span className="InsightLegendButton-title">{filters.show_legend ? 'Hide' : 'Show'} legend</span>
-        </Button>
-    ) : null
-}
 
 function shouldHighlightThisRow(
     hiddenLegendKeys: Record<string, boolean | undefined>,
@@ -58,69 +38,6 @@ function shouldHighlightThisRow(
     ).length
     const isSkipped = hiddenLegendKeys[rowIndex]
     return highlightedSeries !== null && !isSkipped && highlightedSeries + numberOfSeriesToSkip === rowIndex
-}
-
-function InsightLegendRow({
-    hiddenLegendKeys,
-    rowIndex,
-    item,
-    hasMultipleSeries,
-    toggleVisibility,
-    filters,
-    highlighted,
-}: {
-    hiddenLegendKeys: Record<string, boolean | undefined>
-    rowIndex: number
-    item: IndexedTrendResult
-    hasMultipleSeries: boolean
-    toggleVisibility: (index: number) => void
-    filters: Partial<FilterType>
-    highlighted: boolean
-}): JSX.Element {
-    const highlightStyle: Record<string, any> = highlighted
-        ? {
-              style: { backgroundColor: getSeriesColor(item.seriesIndex, false, true) },
-          }
-        : {}
-
-    const rowRef = useRef<HTMLDivElement>(null)
-    useEffect(() => {
-        if (highlighted && rowRef.current) {
-            rowRef.current.scrollIntoView({ block: 'nearest', behavior: 'smooth' })
-        }
-    }, [highlighted])
-
-    const compare = isTrendsFilter(filters) && !!filters.compare
-
-    return (
-        <div key={item.id} className="InsightLegendMenu-item p-2 flex flex-row" ref={rowRef} {...highlightStyle}>
-            <div className="grow">
-                <LemonCheckbox
-                    className="text-xs mr-4"
-                    color={getSeriesColor(item.seriesIndex, compare)}
-                    checked={!hiddenLegendKeys[rowIndex]}
-                    onChange={() => toggleVisibility(rowIndex)}
-                    fullWidth
-                    label={
-                        <InsightLabel
-                            key={item.id}
-                            seriesColor={getSeriesColor(item.seriesIndex, compare)}
-                            action={item.action}
-                            fallbackName={item.breakdown_value === '' ? 'None' : item.label}
-                            hasMultipleSeries={hasMultipleSeries}
-                            breakdownValue={item.breakdown_value === '' ? 'None' : item.breakdown_value?.toString()}
-                            compareValue={compare ? formatCompareLabel(item) : undefined}
-                            pillMidEllipsis={item?.filter?.breakdown === '$current_url'} // TODO: define set of breakdown values that would benefit from mid ellipsis truncation
-                            hideIcon
-                        />
-                    }
-                />
-            </div>
-            {isTrendsFilter(filters) && filters.display === ChartDisplayType.ActionsPie && (
-                <div className="text-muted grow-0">{formatAggregationAxisValue(filters, item.aggregated_value)}</div>
-            )}
-        </div>
-    )
 }
 
 export function InsightLegend({ horizontal, inCardView, readOnly = false }: InsightLegendProps): JSX.Element | null {
