+++ frontend/src/scenes/insights/InsightContainer.tsx
@@ -20,7 +20,8 @@ import {
 import { funnelLogic } from 'scenes/funnels/funnelLogic'
 import clsx from 'clsx'
 import { PathCanvasLabel } from 'scenes/paths/PathsLabel'
-import { InsightLegend, InsightLegendButton } from 'lib/components/InsightLegend/InsightLegend'
+import { InsightLegend } from 'lib/components/InsightLegend/InsightLegend'
+import { InsightLegendButton } from 'lib/components/InsightLegend/InsightLegendButton'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
 import { FunnelStepsTable } from './views/Funnels/FunnelStepsTable'
 import { Animation } from 'lib/components/Animation/Animation'
